requirements.txt
git+https://github.com/Rapptz/discord.py
discord.py
python-dotenv
